backend/requirements.txt
prophet
numpy>=2.0.0
pydantic
fastapi
uvicorn
google-generativeai
supabase
python-dotenv
requests
pandas
scikit-learn
plotly
